requirements.txt
langchain-groq
langgraph
langchain
pydantic
python-dotenv
PyGithub
PyPDF2
python-docx
streamlit
pandas
plotly
pdfminer.six